+++ EasyReminderTDDUITests/AccessibillityIds/AccessibillityIds.swift
@@ -1,0 +1,27 @@
+//
+//  AccessibillityIds.swift
+//  EasyReminderTDDUITests
+//
+//  Created by pozhidaev on 07.06.2023.
+//  Copyright © 2023 Mariya. All rights reserved.
+//
+
+import Foundation
+
+public enum AccessibilityIDs {
+    public enum Main {
+        public static let addButton = "AddId"
+        public static let showClosedButton = "ShowClosedButtonId"
+        public static let checkBox = "CheckId"
+        public static let datelabel = "DateLabelId"
+    }
+    
+    public enum AddNewTask {
+        public static let title = "TitleId"
+        public static let location = "LocationId"
+        public static let description = "description"
+        public static let date = "DateId"
+        public static let myDate = "myDate"
+        public static let priority = "priority"
+    }
+}
